% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_track}
\alias{dplyr_track}
\alias{select.track}
\alias{rename.track}
\alias{filter.track}
\alias{arrange.track}
\alias{mutate.track}
\alias{transmute.track}
\alias{summarise.track}
\alias{summarize.track}
\alias{group_by.track}
\alias{ungroup.track}
\alias{sample_n.track}
\alias{sample_frac.track}
\alias{do.track}
\title{Dplyr Methods For Track Tables}
\usage{
select.track(.data, ...)

rename.track(.data, ...)

filter.track(.data, ...)

arrange.track(.data, ...)

mutate.track(.data, ...)

transmute.track(.data, ...)

summarise.track(.data, ...)

summarize.track(.data, ...)

group_by.track(.data, ...)

ungroup.track(.data, ...)

sample_n.track(.data, ...)

sample_frac.track(.data, ...)

do.track(.data, ...)
}
\arguments{
\item{.data}{A track table.}

\item{...}{Additional arguments to be passed to the corresponding
\code{\link[dplyr]{dplyr}} method.}
}
\description{
\code{\link[dplyr]{dplyr}} methods for track tables objects.
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}},
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}},
 \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}},
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}},
 \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{sample_frac}},
 \code{\link[dplyr]{do}}
}
