% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_track_df}
\alias{dplyr_track_df}
\alias{select.track_df}
\alias{rename.track_df}
\alias{filter.track_df}
\alias{arrange.track_df}
\alias{mutate.track_df}
\alias{transmute.track_df}
\alias{summarise.track_df}
\alias{summarize.track_df}
\alias{group_by.track_df}
\alias{ungroup.track_df}
\alias{sample_n.track_df}
\alias{sample_frac.track_df}
\alias{do.track_df}
\title{Dplyr Methods For Track Tables}
\usage{
select.track_df(.data, ...)

rename.track_df(.data, ...)

filter.track_df(.data, ...)

arrange.track_df(.data, ...)

mutate.track_df(.data, ...)

transmute.track_df(.data, ...)

summarise.track_df(.data, ...)

summarize.track_df(.data, ...)

group_by.track_df(.data, ...)

ungroup.track_df(.data, ...)

sample_n.track_df(.data, ...)

sample_frac.track_df(.data, ...)

do.track_df(.data, ...)
}
\arguments{
\item{.data}{A track table.}

\item{...}{Additional arguments to be passed to the corresponding
\code{\link[dplyr]{dplyr}} method.}
}
\description{
\code{\link[dplyr]{dplyr}} methods for track tables objects.
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}},
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}},
 \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}},
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}},
 \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{sample_frac}},
 \code{\link[dplyr]{do}}
}
