% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{filter}
\alias{filter}
\title{Subset rows using column values}
\arguments{
\item{.data}{See \code{\link[dplyr]{filter}} for an explanation.}

\item{...}{Other args}
}
\description{
This is a re-export of the \code{\link[dplyr]{filter}} function
 in the \code{\link[dplyr]{dplyr}} package. See the help
}
