% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{is_track}
\alias{is_track}
\title{Check Validity of Track Table}
\usage{
is_track(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical indicating whether the variable contains a track table
 (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Test whether a variable contains a track table as produced
 by \code{\link{track_df}}, \code{\link{track_tbl}}, or \code{\link{track_dt}}.
}
\examples{
data(tracks)

is_track(tracks)

}
\seealso{
\code{\link{track_df}}, \code{\link{track_tbl}}, \code{\link{track_dt}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
