% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_df.R
\name{is_track_df}
\alias{is_track_df}
\title{Check Validity of Track Table}
\usage{
is_track_df(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical indicating whether the variable contains a track table
 (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Test whether a variable contains a track table as produced
 by \code{\link{track_df}}.
}
\examples{
# TODO

}
\seealso{
\code{\link{track_df}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
