% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_dims}
\alias{n_dims}
\title{Number of Spatial Dimensions of a Track Table}
\usage{
n_dims(x)
}
\arguments{
\item{x}{A track data table as produced by \code{\link{track_df}}.}
}
\value{
A numeric value.
}
\description{
Track tables produced by \code{\link{track_df}} can have 2 (x,y)
 or 3 (x, y, z) spatial dimensions. This function returns the number of
 spatial dimensions of a track table.
}
\examples{
data(tracks)

n_dims(tracks)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
