% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projection}
\alias{projection}
\alias{projection<-}
\alias{project}
\title{Access/Modify the Projection of a Track Table}
\usage{
projection(x)

projection(x) <- value

project(x, value)
}
\arguments{
\item{x}{A track table.}

\item{value}{A character string or a \code{\link[sp:CRS]{sp::CRS}} object
representing the projection of the coordinates. \code{"+proj=longlat"} is
suitable for the outputs of most GPS trackers.}
}
\value{
A track table.
}
\description{
Functions to access or modify the projection of a data table.
 Changing the projection will trigger automatically the conversion of the
 locations in the new coordinate system.
}
\note{
It is not possible to modify the projection if missing coordinates
 are present.
}
\examples{
library(trackdf)
t_df <- track_df(x = tracks$x, y = tracks$y, t = tracks$t, id = tracks$id,
  proj = "+proj=longlat", tz = "Africa/Windhoek")

projection(t_df)
t_df_projected <- project(t_df, "+proj=somerc")
projection(t_df_projected)
projection(t_df_projected) <- "+proj=longlat"
projection(t_df_projected)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
