% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_df.R
\name{[.track_df}
\alias{[.track_df}
\alias{[<-.track_df}
\title{Extract or Replace Parts of a Track Table}
\usage{
\method{[}{track_df}(x, i, j, drop = TRUE)

\method{[}{track_df}(x, i, j) <- value
}
\arguments{
\item{x}{A track table.}

\item{i, j}{Row and column indices. If \code{j} is omitted, \code{i} is used as
column index.}

\item{drop}{Coerce to a vector if fetching one column via \code{track_df[, j]}.
Default \code{FALSE}, ignored when accessing a column via \code{track_df[j]}.}

\item{value}{A value to be assigned to x.}
}
\description{
Accessing columns, rows, or cells via $, [[, or [ is mostly
 similar to regular \code{\link[base:Extract.data.frame]{data frames}}.
 However, the behavior is different for track tables and data frames in some
 cases, because track tables inherit from \code{\link[tibble]{tibble}}:
 \itemize{
  \item \code{[} always returns a track table by default, even if only one
   column is accessed.
  \item Partial matching of column names with $ and [[ is not supported, a
   warning is given and NULL is returned.
 }
 For more info, refer to \link[tibble:subsetting]{tibble's subsetting
 documentation}
}
\examples{
# TODO

}
\seealso{
\code{\link{track_df}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
