% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{t_duplicates}
\alias{t_duplicates}
\title{Find Duplicated Timestamps in a Track Table}
\usage{
t_duplicates(x)
}
\arguments{
\item{x}{\code{\link{track}} table as produced by the \code{\link{track}}
function. The table should only contain the observations for a single
individual. See the examples below to handle multiple individuals.}
}
\value{
A vector of timestamps duplicated in \code{x}. If no duplicated
 timestamps are detected, the function returns an empty vector.
}
\description{
This function attempts to automatically detect duplicated
 timestamps in \code{\link{track}} tables.
}
\note{
Missing timestamps ("NA") are ignored.
}
\examples{
# Create data set with duplicated data
data(short_tracks)
t_df <- track(x = short_tracks$x, y = short_tracks$y, t = short_tracks$t,
              id = short_tracks$id, proj = "+proj=longlat",
              tz = "Africa/Windhoek", table = "df")
t_df <- bind_tracks(t_df, t_df[1:10, ], t_df[100:110, ])

# Find missing data for one individual
duplicated_one <- t_duplicates(t_df[t_df$id == 1, ])

# Find missing data for all individuals
duplicated_all <- lapply(split(t_df, t_df$id), t_duplicates)

}
\seealso{
\code{\link{track}}, \code{\link{fix_t_duplicates}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
