% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_df.R
\name{track_df}
\alias{track_df}
\title{Build a Track Table}
\usage{
track_df(x, y, z, t, id, ..., proj, origin, period, tz)
}
\arguments{
\item{x, y, z}{Numaric vectors representing the coordinates of the locations.
\code{x} and \code{y} are required. \code{z} can be ignored if the
trajectories are 2-dimensional. Note: if the vectors are not of the same
length, the shorter ones will be recycled to match the length of the longer
one.}

\item{t}{A numeric vector or a vector of objects that can be coerced to
date-time objects by \code{link[lubridate]{as_datetime}} representing the
times (or frames) at which each location was recorded. If numeric, the
origin and period of the time points can be set using \code{origin} and
\code{period} below.}

\item{id}{A vector that can be coerced to a character vector by
\code{\link{as.character}} representing the identity of the animal to which
each location belong.}

\item{...}{A set of name-value pairs. Arguments are evaluated sequentially,
so you can refer to previously created elements. These arguments are
processed with \code{\link[rlang:quotation]{rlang::quos()}} and support
unquote via \code{!!} and unquote-splice via \code{!!!}. Use \code{:=} to
create columns that start with a dot.}

\item{proj}{A character string or a \code{\link[sp:CRS]{sp::CRS}} object
representing the projection of the coordinates. Leave empty if the
coordinates are not projected (e.g., output of video tracking).
\code{"+proj=longlat"} is suitable for the outputs of most GPS trackers.}

\item{origin}{Something that can be coerced to a date-time object by
\code{link[lubridate]{as_datetime}} representing the start date and time of
the observations when \code{t} is a numeric vector.}

\item{period}{A character vector in a shorthand format (e.g. "1 second") or
ISO 8601 specification. This is used when \code{t} is a numeric vector to
represent time unit of the observations. All unambiguous name units and
abbreviations are supported, "m" stands for months, "M" for minutes unless
ISO 8601 "P" modifier is present (see examples). Fractional units are
supported but the fractional part is always converted to seconds. See
\code{\link[lubridate]{period}} for more details.}

\item{tz}{A time zone name. See \code{\link{OlsonNames}}.}
}
\value{
A track table, which is a colloquial term for an object of class
 \code{track_df}. A \code{track_df} object is also a tibble and a data frame,
 i.e. it has class \code{\link[tibble]{tbl_df}} and \code{\link{data.frame}}.
}
\description{
\code{track_df} constructs a track table. It is used like
 \code{\link[tibble]{tibble}}, but with a notable difference: it has an extra
 attributes to store the projection of the track coordinates, and modifying
 the projection will automatically trigger the appropriate conversion of the
 coordinates.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
