% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{xy_errors}
\alias{xy_errors}
\title{Find Inconsistent Locations in a Track Table}
\usage{
xy_errors(x, s = 15, span = 0.05, degree = 2, ...)
}
\arguments{
\item{x}{\code{\link{track}} table as produced by the \code{\link{track}}
function. The table should only contain the observations for a single
individual. See the examples below to handle multiple individuals.}

\item{s}{The discrimination threshold of the outlier detection algorithm.
Higher values correspond to less outliers.}
}
\value{
A vector of indices for the inconsistent locations in \code{x}. If no
 inconsistencies are detected, the function returns an empty vector.
}
\description{
This function attempts to automatically detect inconsistent
 locations (for instance due to a writing error or GPS innacuracies) in
 \code{\link{track}} tables.
}
\examples{
# TODO

}
\seealso{
\code{\link{track}}, \code{\link{fix_xy_errors}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
