% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{conversions}
\alias{conversions}
\alias{as_track_df}
\alias{as_track_df.MoveStack}
\alias{as_track_df.Move}
\alias{as_move}
\alias{as_move.track_df}
\title{Convert a Track Table to Other Formats}
\usage{
as_track_df(x, ...)

\method{as_track_df}{MoveStack}(x, ...)

\method{as_track_df}{Move}(x, id, ...)

as_move(x, ...)

\method{as_move}{track_df}(x, ...)
}
\arguments{
\item{x}{A track table to convert.}

\item{...}{Other parameters to be passed to:
\itemize{
  \item{\code{\link{track_df}} if `as_track_df` is used.}
  \item{\code{\link[moveVis:df2move]{moveVis::df2move}} if `as_move` is used.}
}}

\item{id}{A character vector representing the identity of the animal to which
each location belong.}
}
\value{
The coordinates converted in the chosen format.
}
\description{
The following methods will convert track tables to other
 common formats used for processing tracking and spatial data.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
